test(control-flow): isolate fault recovery fixture
